Panic on db connection errors in tests
I had changed this because initially, I had felt this
would be a more graceful way of failing, however there
is a very long timeout on connections, and this ends
up causing all of the tests to take exceedingly long
just to fail because, in 99.9% of cases, mongo just
needs to be started. I came to the conclusion that
it just wasn't worth all of the extra line noise.