Collapse all methods into one
Unfortunately, there doesn't seem to be a good way to
have a slice of structs be seen as equal to a slice of
interfaces implemented by that struct. Because of that
each individual controller will need to implement their
own 'All' method. Given they're only 3 lines of real
code, it's not the worst thing, but a bit annoying.